El CRUD ya está completo

+++ src/app/Tareas/list/list.component.ts
@@ -3,6 +3,7 @@ import {TareaServiceService} from '../../Service/tarea-service.service';
 import { Router} from '@angular/router';
 import {Tarea} from '../../Models/Tarea';
 import {trigger, style, animate, transition, state} from '@angular/animations';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list',
@@ -40,6 +41,10 @@ export class ListComponent implements OnInit {
     this.router.navigate(["create"]);
   }
 
+  edit(id: number){
+    this.router.navigate(["tarea", id]);
+  }
+
   reloadData() {
     this.service.getTareas()
     .subscribe(data=>{
@@ -50,12 +55,29 @@ export class ListComponent implements OnInit {
   }
 
   deleteTodo(id: number) {
-    this.service.delete(id)
-      .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log(error));
+    Swal.fire({
+      title: '¿Deseas borrar esta tarea?',
+      text: "Será borrada definitivamente",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#5689A3',
+      cancelButtonColor: '#8F2B45',
+      confirmButtonText: 'Sí'
+      }).then((result) => {
+      if (result.value) {
+        this.service.delete(id)
+        .subscribe(
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+        Swal.fire(
+          'Borrado',
+          'La tarea ha sido eliminada',
+          'success'
+        )
+      }
+    })
   }
 }
